fix "issues" link in README and update CLA part
the link for report issues points to the wrong repository, I hadn't
noticed that before.

I also updated the part related to the CLA and removed a line that seems
to me "duplicate", in any case now CLA assistant write to the
contributor is CLA is not signed and explain what to do.